+++ libraries/base/changelog.md
@@ -30,6 +30,8 @@
   * Add `iterate'`, a strict version of `iterate`, to `Data.List`
     and `Data.OldList` (#3474)
 
+  * Add `Data` instances for `IntPtr` and `WordPtr` (#13115)
+
 ## 4.10.0.0 *April 2017*
   * Bundled with GHC *TBA*